+++ R/wordbankr.R
@@ -182,7 +182,10 @@ factor_demographics <- function(admins) {
       sex = factor(.data$sex, levels = c("Female", "Male", "Other")),
       ethnicity = factor(.data$ethnicity,
                          levels = c("Hispanic", "Non-Hispanic")),
-      race = factor(.data$race, levels = c("Asian", "Black", "Other/Mixed", "White")),
+      # dataset versions before v2.1 label this level "Other"
+      race = factor(dplyr::if_else(.data$race == "Other", "Other/Mixed",
+                                   .data$race),
+                    levels = c("Asian", "Black", "Other/Mixed", "White")),
       birth_order = factor(.data$birth_order,
                            levels = c("First", "Second", "Third", "Fourth",
                                       "Fifth", "Sixth", "Seventh", "Eighth",